done with adding security to auth flow and journals

+++ src/main/java/com/edigest/journalApp/service/JournalEntryService.java
@@ -24,13 +24,19 @@ public class JournalEntryService {
 
     @Transactional
     public void saveEntry(JournalEntry journalEntry, String userName){
-        User user = userService.findByUserName(userName);
-        journalEntry.setDate(LocalDateTime.now());
-        JournalEntry saved = journalEntryRepository.save(journalEntry);
-        user.getJournalEntries().add(saved);
+        try {
+            User user = userService.findByUserName(userName);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
 
-        userService.saveEntry(user);
-    }
+            userService.saveUser(user);
+
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("an error occurred while saving te entry.",e);
+        }
+        }
 
      public void saveEntry(JournalEntry journalEntry){
         journalEntryRepository.save(journalEntry);
@@ -46,11 +52,26 @@ public class JournalEntryService {
         
     }
 
-    public void deleteByID(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x-> x.getId().equals(id));
-        userService.saveEntry(user);
-          journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteByID(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+          User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+          if(removed){
+              userService.saveUser(user);
+              journalEntryRepository.deleteById(id);
+          }
+      }catch (Exception e){
+          System.out.println(e);
+          throw new RuntimeException("An error occurred while deleting the entry,",e);
+      }
+      return removed;
+
     }
+
+//    public List<JournalEntry> findByUserName(String userName){
+////        findByUserName(us)
+//    }
 
 }
